config variable package is now called bs_package
i'm going to adopt the bs_ prefix for all the config
variables.  the original reasoning for having (or not)
the prefix, that is, whether the value is BS-specific
or not (package is not as this value is used in the
upstream url and tarball) does not look so strong anymore.

i think it makes more sense to repurpose the prefix for
bs-update in general.  the other config vars will follow.