tests for post-option-processing code w/o clutter
basically, the first half of bs-update.in has
bsu_configuration_processing=1, the rest of the code is
under bsu_configuration_processing=0.

the asymmetry in tests/fakes:test irks me, but it's not that much
of a deal as all tests for the option/argument-handling code use
BSU_QUIET_FAKE=1 already/anyway.