WIP: breaking change in bsu_tarball_hook behavior
bs-update cannot repack all tarballs, repacking breaks
checksums/signatures.  this feature is in fact an idiosyncracy:
i wrote it for the special case of the packager being the
upstream maintainer as well.

bsu_tarball_hook() now must repack the tarball on its own: it's
run in a directory containing the tarball, with its name in $1.
possible implementation looks very much like the late repack():

  bsu_tarball_hook()
  (
    set -e
    local tarball=${1?}
    local dir=${tarball%.tar.gz}
    mkdir $dir
    tar -xzf $tarball --strip-components=1 -C $dir
    cd $dir
    do_stuff
    tar -czf $tarball $dir
    rm -rf $dir
  )