BC-breaking change in bsu_tarball_hook behavior
bs-update cannot repack all tarballs, it breaks checksums and
signatures.  this feature is in fact an idiosyncracy written
for the original use case (now a special case) of the packager
being the upstream maintainer as well.

bsu_tarball_hook() now must repack the tarball on its own if it
wants that to happen: the hook is run in a directory containing
the tarball, with its name in $1.  possible implementation looks
very much like the late repack():

  bsu_tarball_hook()
  (
    set -e
    local tarball=${1?}
    local dir=${tarball%.tar.gz}
    mkdir $dir
    tar -xzf $tarball --strip-components=1 -C $dir
    cd $dir
    # --- >8 ---
    # massage it
    # --- 8< ---
    tar -czf $tarball $dir
    rm -rf $dir
  )